Ecoinvent (#451)
- Ecoinvent import from simapro csv
- Copy-pasted 4 EI processes from simapro
- Moved import functions in the import_ module
- Removed materialAndSpinningProcessUuid in the data
