+++ data/import_ecoinvent.py
@@ -2,41 +2,14 @@
 
 from bw2data.project import projects
 from common.import_ import add_missing_substances
-from zipfile import ZipFile
+from import_agribalyse import import_simapro_csv
 import bw2data
 import bw2io
-import os
-import shutil
 
 PROJECT = "textile"
 # Ecoinvent
-DATAPATH = "./ECOINVENT3.9.1.zip"
-DBNAME = "Ecoinvent 3.9.1"
+DATAPATH = "./Ecoinvent3.9.1.CSV.zip"
 BIOSPHERE = "biosphere3"
-
-
-def import_ecoinvent(datapath=DATAPATH, project=PROJECT, dbname=DBNAME):
-    """
-    Import file at path `datapath` into database named `dbname` in the project
-    """
-    projects.set_current(project)
-    # projects.create_project(project, activate=True, exist_ok=True)
-
-    # unzip
-    with ZipFile(datapath) as zf:
-        print("### Extracting the zip file...")
-        zf.extractall()
-        unzipped = datapath[0:-4]
-
-    print(f"### Importing {dbname} database from {unzipped}...")
-    ecoinvent = bw2io.importers.SingleOutputEcospold2Importer(
-        os.path.join(unzipped, "datasets"), dbname
-    )
-    shutil.rmtree(unzipped)
-    ecoinvent.apply_strategies()
-    ecoinvent.add_unlinked_flows_to_biosphere_database()
-    ecoinvent.write_database()
-    print(f"### Finished importing {dbname}")
 
 
 def main():
@@ -46,11 +19,10 @@ def main():
     bw2io.bw2setup()
     add_missing_substances(PROJECT, BIOSPHERE)
 
-    # Import Ecoinvent
-    if DBNAME not in bw2data.databases:
-        import_ecoinvent()
+    if (db := "Ecoinvent 3.9.1") not in bw2data.databases:
+        import_simapro_csv(DATAPATH, db)
     else:
-        print(f"### {DBNAME} is already imported")
+        print(f"{db} already imported")
 
 
 if __name__ == "__main__":
