+++ public/data/countries.json
@@ -53,7 +53,7 @@
     "code": "CN",
     "name": "Chine",
     "zone": "Asia",
-    "electricityProcessUuid": "8f923f3d-0bd2-4326-99e2-f984b4454226",
+    "electricityProcessUuid": "a21f671eed68dd9e7d4694f8cf1b28df",
     "heatProcessUuid": "2e8de6f6-0ea1-455b-adce-ea74d307d222",
     "airTransportRatio": 0.33,
     "scopes": ["food", "textile"],
@@ -113,7 +113,7 @@
     "code": "FR",
     "name": "France",
     "zone": "Europe",
-    "electricityProcessUuid": "05585055-9742-4fff-81ff-ad2e30e1b791",
+    "electricityProcessUuid": "bae4d1067750357f0217b4273f7a51ff",
     "heatProcessUuid": "12fc43f2-a007-423b-a619-619d725793ea",
     "airTransportRatio": 0,
     "scopes": ["food", "textile"],
