+++ public/data/food/processes.json
@@ -490,25 +490,25 @@
     "comment": "",
     "database": "Ecobalyse",
     "impacts": {
-      "acd": 0.2342054931,
-      "ozd": 2.927223966e-7,
-      "cch": 16.36861768,
-      "fwe": 0.0006655837445,
-      "swe": 0.0384158816,
-      "tre": 1.044794298,
-      "pco": 0.02512768195,
-      "pma": 1.560351933e-6,
-      "ior": 1.029909812,
-      "fru": 40.79065872,
-      "mru": 1.421252275e-5,
-      "ldu": 1365.685376,
-      "wtu": 1.852505802,
-      "htc": 7.016937169e-9,
-      "htn": 2.0015754e-7,
-      "etf": 22.789406469,
-      "etf-c": 24.33562694582,
-      "htc-c": 1.4659670784e-9,
-      "htn-c": 4.878298476e-8
+      "acd": 0.3963934994,
+      "ozd": 4.733772908e-7,
+      "cch": 26.4068562,
+      "fwe": 0.001103831682,
+      "swe": 0.06590072544,
+      "tre": 1.766745681,
+      "pco": 0.03956660902,
+      "pma": 2.644178814e-6,
+      "ior": 1.205204885,
+      "fru": 55.79484487,
+      "mru": 2.826105362e-5,
+      "ldu": 2173.669941,
+      "wtu": 2.987802249,
+      "htc": 1.065153857e-8,
+      "htn": 2.966083703e-7,
+      "etf": 34.724851688,
+      "etf-c": 37.173352674300006,
+      "htc-c": 2.266844818e-9,
+      "htn-c": 7.347609014e-8
     }
   },
   {
@@ -4266,25 +4266,25 @@
     "comment": "This dataset includes packaging",
     "database": "Ecobalyse",
     "impacts": {
-      "acd": 0.1906283406,
-      "ozd": 2.296659173e-7,
-      "cch": 13.25510547,
-      "fwe": 0.000524402892,
-      "swe": 0.03125502438,
-      "tre": 0.8514034208,
-      "pco": 0.02025696048,
-      "pma": 1.268118298e-6,
-      "ior": 0.3723431438,
-      "fru": 22.82640731,
-      "mru": 1.069879544e-5,
-      "ldu": 1112.640029,
-      "wtu": 1.453509676,
-      "htc": 5.6419722e-9,
-      "htn": 1.622825857e-7,
-      "etf": 19.150229125,
-      "etf-c": 20.358304961720002,
-      "htc-c": 1.1301370344e-9,
-      "htn-c": 3.982308748e-8
+      "acd": 0.3228294878,
+      "ozd": 3.769196188e-7,
+      "cch": 21.4373791,
+      "fwe": 0.000881623388,
+      "swe": 0.05365820922,
+      "tre": 1.439873575,
+      "pco": 0.03202628157,
+      "pma": 2.151556971e-6,
+      "ior": 0.5152279986,
+      "fru": 35.05647701,
+      "mru": 2.214990048e-5,
+      "ldu": 1771.219616,
+      "wtu": 2.378901732,
+      "htc": 8.604573969e-9,
+      "htn": 2.409006703e-7,
+      "etf": 28.878935865,
+      "etf-c": 30.82247002048,
+      "htc-c": 1.7829408912e-9,
+      "htn-c": 5.995069698e-8
     }
   },
   {
