+++ public/data/textile/processes.json
@@ -1,32 +1,32 @@
 [
   {
     "search": "",
-    "name": "Mix électrique réseau, CN",
+    "name": "(EI) market group for electricity, medium voltage, CN",
     "info": "Energie > Electricité > Mix moyen",
     "unit": "kWh",
-    "source": "Base Impacts 2.01",
-    "uuid": "8f923f3d-0bd2-4326-99e2-f984b4454226",
+    "source": "Ecoinvent 3.9.1",
+    "uuid": "a21f671eed68dd9e7d4694f8cf1b28df",
     "impacts": {
-      "acd": 0.00896069,
+      "acd": 0.0015271008,
       "bvi": 0.0,
-      "cch": 1.05738,
-      "etf": 0.0,
-      "etf-c": 0,
-      "fru": 10.6736,
-      "fwe": 2.21258e-8,
-      "htc": 0.0,
-      "htc-c": 0,
-      "htn": 0.0,
-      "htn-c": 0,
-      "ior": 0.0144336,
-      "ldu": 0.0,
-      "mru": 8.51552e-8,
-      "ozd": 6.31189e-13,
-      "pco": 0.00649117,
-      "pma": 3.32057e-7,
-      "swe": 0.00235429,
-      "tre": 0.0256924,
-      "wtu": 0.0
+      "cch": 0.28407856,
+      "etf": 0.711914855,
+      "etf-c": 0.71663734332,
+      "fru": 2.5963146,
+      "fwe": 5.2966609e-5,
+      "htc": 8.19e-12,
+      "htc-c": 1.63813654e-11,
+      "htn": 8.39e-11,
+      "htn-c": 1.5516067e-10,
+      "ior": 0.0048357578,
+      "ldu": 0.5567515,
+      "mru": 2.63e-7,
+      "ozd": 1.54e-9,
+      "pco": 0.00090117771,
+      "pma": 2.16e-8,
+      "swe": 0.00032668138,
+      "tre": 0.0034795272,
+      "wtu": 0.030280276
     },
     "heat_MJ": 0.0,
     "elec_pppm": 0.0,
@@ -34,7 +34,7 @@
     "waste": 0.0,
     "alias": null,
     "step_usage": "Energie",
-    "correctif": "Correction de l'indicateur radiations ionisantes"
+    "correctif": ""
   },
   {
     "search": "",
@@ -75,32 +75,32 @@
   },
   {
     "search": "",
-    "name": "Mix électrique réseau, FR",
+    "name": "(EI) market for electricity, medium voltage, FR",
     "info": "Energie > Electricité > Mix moyen",
     "unit": "kWh",
-    "source": "Base Impacts 2.01",
-    "uuid": "05585055-9742-4fff-81ff-ad2e30e1b791",
+    "source": "Ecoinvent 3.9.1",
+    "uuid": "bae4d1067750357f0217b4273f7a51ff",
     "impacts": {
-      "acd": 0.000209809,
+      "acd": 9.2167346e-5,
       "bvi": 0.0,
-      "cch": 0.0813225,
-      "etf": 0.0,
-      "etf-c": 0,
-      "fru": 9.31353,
-      "fwe": 3.26897e-8,
-      "htc": 0.0,
-      "htc-c": 0,
-      "htn": 0.0,
-      "htn-c": 0,
-      "ior": 0.718762,
-      "ldu": 0.0,
-      "mru": 4.85798e-8,
-      "ozd": 8.77128e-10,
-      "pco": 0.000210846,
-      "pma": 4.15844e-9,
-      "swe": 7.61251e-5,
-      "tre": 0.00049761,
-      "wtu": 0.0
+      "cch": 0.021581196,
+      "etf": 0.09126594,
+      "etf-c": 0.1256067558,
+      "fru": 3.2188003,
+      "fwe": 4.2592527e-6,
+      "htc": 6.4e-12,
+      "htc-c": 1.27903874e-11,
+      "htn": 1.19e-11,
+      "htn-c": 2.320931e-11,
+      "ior": 0.14628567,
+      "ldu": 0.096699359,
+      "mru": 1.75e-7,
+      "ozd": 8.72e-10,
+      "pco": 5.051977e-5,
+      "pma": 1.25e-9,
+      "swe": 2.7334921e-5,
+      "tre": 0.0002017688,
+      "wtu": 0.0060828454
     },
     "heat_MJ": 0.0,
     "elec_pppm": 0.0,
@@ -108,7 +108,7 @@
     "waste": 0.0,
     "alias": null,
     "step_usage": "Energie",
-    "correctif": "Correction de l'indicateur radiations ionisantes"
+    "correctif": ""
   },
   {
     "search": "",
@@ -925,6 +925,53 @@
     "correctif": "Ajout données \"EP&L Kering\" pour la consommation d'eau"
   },
   {
+    "name": "(EI) polyester fibre production, finished",
+    "info": "Textile > Matières > Matières synthétiques",
+    "unit": "kg",
+    "source": "Ecoinvent 3.9.1",
+    "uuid": "91f1b47ef55e48fab5b5e2ae8ab7bbef",
+    "impacts": {
+      "acd": 0.0186312,
+      "bvi": 0.0,
+      "cch": 4.1593321,
+      "etf": 12.67545,
+      "etf-c": 13.777083034,
+      "fru": 86.374758,
+      "fwe": 0.0010940637,
+      "htc": 5.23e-10,
+      "htc-c": 1.0454992e-9,
+      "htn": 5.04e-9,
+      "htn-c": 9.956801e-9,
+      "ior": 0.28591251,
+      "ldu": 9.6837138,
+      "mru": 3.88e-5,
+      "ozd": 1.55e-5,
+      "pco": 0.018726196,
+      "pma": 1.92e-7,
+      "swe": 0.0040132771,
+      "tre": 0.036363971,
+      "wtu": 2.1062805
+    },
+    "heat_MJ": 0.0,
+    "elec_pppm": 0.0,
+    "elec_MJ": 0.0,
+    "waste": 0.0319569,
+    "alias": null,
+    "step_usage": "Matières, Filature",
+    "correctif": "Ajout données \"EP&L Kering\" pour la consommation d'eau",
+    "id": "ei-pet",
+    "materialProcessUuid": "4dad2986-ce83-4960-6a06-e9ab85a0bcc1",
+    "recycledProcessUuid": "3e1c26d3-23ef-323e-e848-f808f54d35bf",
+    "recycledFrom": null,
+    "shortName": "(EI) Polyester",
+    "origin": "Synthetic",
+    "primary": true,
+    "geographicOrigin": "Asie - Pacifique",
+    "defaultCountry": "CN",
+    "priority": 80,
+    "cff": null
+  },
+  {
     "search": "",
     "name": "Filament de polyester, inventaire partiellement agrégé",
     "info": "Textile > Matières > Matières synthétiques",
@@ -1737,6 +1784,54 @@
     "alias": null,
     "step_usage": "Matières, Filature",
     "correctif": "Ajout données \"EP&L Kering\" pour la consommation d'eau"
+  },
+  {
+    "search": "",
+    "name": "(EI) fibre production, cotton, ginning, RoW",
+    "info": "Textile > Matières > Matières naturelles",
+    "unit": "kg",
+    "source": "Ecoinvent 3.9.1",
+    "uuid": "23d8981e4e4f17b44cd5299eef7784ea",
+    "impacts": {
+      "acd": 0.050004554,
+      "bvi": 0.0,
+      "cch": 2.758903,
+      "etf": 296.78549,
+      "etf-c": 580.72593,
+      "fru": 21.091752,
+      "fwe": 0.0020965554,
+      "htc": 1.07e-9,
+      "htc-c": 2.1392048e-9,
+      "htn": 3.79e-8,
+      "htn-c": 7.5609412e-8,
+      "ior": 0.11439048,
+      "ldu": 206.8445,
+      "mru": 2.45e-5,
+      "ozd": 2.1e-7,
+      "pco": 0.013716303,
+      "pma": 3.35e-7,
+      "swe": 0.046469433,
+      "tre": 0.21154201,
+      "wtu": 32.330333
+    },
+    "heat_MJ": 0.0,
+    "elec_pppm": 0.0,
+    "elec_MJ": 0.0,
+    "waste": 0.201201,
+    "alias": null,
+    "step_usage": "Matières, Filature",
+    "correctif": "Ajout données \"EP&L Kering\" pour la consommation d'eau",
+    "id": "ei-fil-coton",
+    "materialProcessUuid": "14a03569-d26b-9496-92e5-dfe8cb1855fe",
+    "recycledProcessUuid": "9c6ab710-4a08-c720-cede-24428a013fda",
+    "recycledFrom": null,
+    "shortName": "(EI) Coton",
+    "origin": "NaturalFromVegetal",
+    "primary": true,
+    "geographicOrigin": "Asie - Pacifique",
+    "defaultCountry": "CN",
+    "priority": 100,
+    "cff": null
   },
   {
     "search": "",
